requirements.txt
bcc==0.1.10
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.4
pyroute2==0.9.1
python-iptables==1.0.1
PyYAML==6.0.2
traitlets==5.14.3
traittypes==0.2.1
Werkzeug==3.1.3
